Kimchi: move more berkeley expr related code into berkeley_columns
