+++ kimchi/src/circuits/polynomials/complete_add.rs
@@ -16,8 +16,8 @@
 //~
 use crate::circuits::{
     argument::{Argument, ArgumentEnv, ArgumentType},
-    berkeley_columns::BerkeleyChallengeTerm,
-    expr::{constraints::ExprOps, Cache},
+    berkeley_columns::{constraints::ExprOps, BerkeleyChallengeTerm},
+    expr::Cache,
     gate::{CircuitGate, GateType},
     wires::COLUMNS,
 };
