+++ kimchi/src/circuits/polynomials/generic.rs
@@ -36,8 +36,8 @@
 use crate::{
     circuits::{
         argument::{Argument, ArgumentEnv, ArgumentType},
-        berkeley_columns::BerkeleyChallengeTerm,
-        expr::{constraints::ExprOps, Cache},
+        berkeley_columns::{constraints::ExprOps, BerkeleyChallengeTerm},
+        expr::Cache,
         gate::{CircuitGate, GateType},
         polynomial::COLUMNS,
         wires::GateWires,
