+++ kimchi/src/circuits/polynomials/turshi.rs
@@ -82,9 +82,11 @@ use crate::{
     alphas::Alphas,
     circuits::{
         argument::{Argument, ArgumentEnv, ArgumentType},
-        berkeley_columns::{BerkeleyChallengeTerm, BerkeleyChallenges, Column, E},
+        berkeley_columns::{
+            constraints::ExprOps, BerkeleyChallengeTerm, BerkeleyChallenges, Column, E,
+        },
         constraints::ConstraintSystem,
-        expr::{self, constraints::ExprOps, Cache},
+        expr::{self, Cache},
         gate::{CircuitGate, GateType},
         wires::{GateWires, Wire, COLUMNS},
     },
