+++ kimchi/src/snarky/range_checks.rs
@@ -157,8 +157,8 @@ pub fn range_check<F: PrimeField>(
 #[cfg(test)]
 mod test {
     use crate::{
-        circuits::expr::constraints::ExprOps, loc, snarky::api::SnarkyCircuit, FieldVar, RunState,
-        SnarkyResult,
+        circuits::berkeley_columns::constraints::ExprOps, loc, snarky::api::SnarkyCircuit,
+        FieldVar, RunState, SnarkyResult,
     };
     use mina_curves::pasta::{Fp, Vesta, VestaParameters};
     use mina_poseidon::{
